Move logo to emotion to avoid image resolution
After bumping razzle/webpack/whatever the build would fail since the
images used in logo.css could not be resolved at the time of build.

This could probably be solved by setting up some sort of resolver in our
razzle/webpack config. However this only happens in logo.css and since
we already depend on emotion, i just moved that component to css-in-js
:^)